LayoutTests/imported/w3c/web-platform-tests/mst-content-hint/RTCRtpSendParameters-degradationEffect-expected.txt


FAIL Maintain-framerate reduces resolution on bandwidth cut assert_equals: Expect initial param.degradationPreference to be undefined expected (undefined) undefined but got (string) "balanced"
